requirements.txt
# Core dependencies
huggingface_hub>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0  # Used for Ollama's OpenAI-compatible API

# Web and API dependencies
flask>=2.3.3
flask-cors>=4.0.0

# Interface dependencies
gradio>=4.0.0
# asyncio-mqtt is in your original file, kept for optional MQTT interface
asyncio-mqtt>=0.13.0

# AI Framework dependencies (optional, install as needed)
smolagents>=0.1.0
llama-index>=0.9.0
langgraph>=0.1.0

# Evaluation & Testing dependencies (optional)
ollama>=0.2.0
pandas>=2.0.0
PyYAML>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Development & Formatting dependencies (optional)
black>=23.0.0
flake8>=6.0.0

# Documentation dependencies (optional)
mkdocs>=1.5.0
mkdocs-material>=9.0.0